Fixes bug 210, navigate through main menu via keyboard
